requirements.txt
Django==4.2.1
djangorestframework==3.14.0
Pillow==9.5.0
whitenoise==6.4.0
